app/src/main/java/com/nyt/articles/data/model/Article.java
package com.nyt.articles.data.model;

import android.os.Parcel;
import android.os.Parcelable;

import com.google.gson.annotations.SerializedName;

/**
 * Created by raji on 26/09/2018.
 */
public class Article implements Parcelable {

    @SerializedName("id")
    private long id;

    @SerializedName("title")
    private String title;

    @SerializedName("byline")
    private String byline;

    @SerializedName("published_date")
    private String published_date;

    @SerializedName("url")
    private String url;

    @SerializedName("abstract")
    private String articleAbstract;

    protected Article(Parcel in) {
        id = in.readLong();
        title = in.readString();
        byline = in.readString();
        published_date = in.readString();
        url = in.readString();
        articleAbstract=in.readString();
    }

    public Article() {
    }

    public static final Creator<Article> CREATOR = new Creator<Article>() {
        @Override
        public Article createFromParcel(Parcel in) {
            return new Article(in);
        }

        @Override
        public Article[] newArray(int size) {
            return new Article[size];
        }
    };

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getByline() {
        return byline;
    }

    public void setByline(String byline) {
        this.byline = byline;
    }

    public String getPublished_date() {
        return published_date;
    }

    public void setPublished_date(String published_date) {
        this.published_date = published_date;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }

    public String getArticleAbstract() {
        return articleAbstract;
    }

    public void setArticleAbstract(String articleAbstract) {
        this.articleAbstract = articleAbstract;
    }

    @Override
    public int describeContents() {
        return 0;
    }

    @Override
    public void writeToParcel(Parcel dest, int flags) {
        dest.writeLong(id);
        dest.writeString(title);
        dest.writeString(byline);
        dest.writeString(published_date);
        dest.writeString(url);
        dest.writeString(articleAbstract);
    }
}
